Don't skip the method screen when going back and forth (#477991).
The following UI flow is broken:  OK on keyboard -> Back on method
selection -> OK on keyboard -> method.  In this case, the method screen
is skipped and loader cruises way ahead to searching for the stage2
image on a CD.  Failing that (like, you're doing a PXE boot) it will
display a screen telling you that a CD couldn't be found.  Fix it by
allowing displaying method selection again.