Add a message to install.log when package installation is done (#476953).
install.log is written in-place to /mnt/sysimage/root, while the anaconda.log
is only copied over to the running system at the very end of anaconda.  It is
therefore possible to get an installation that fails during package
installation, but may still look complete on the installed system due to the
presence of an install.log.  So, add a message when it's all done.