Collect DSO deps for NetworkManager plugins.
'cp -a' considered harmful.  jkeating brought to my attention today that
libgio-2.0.so.0 was not getting copying in to the initrd.img which is
required by libnm-settings-plugin-ifcfg-rh.so.

The problem was caused by my patch to mk-images which changed the plugin
instbin line to a 'cp -a' for the *.so files in /usr/$LIBDIR/NetworkManager.
We need to use the instbin function so that get_dso_deps() is run on the
plugin files and we pick up dependent libraries.