btrfs install support
Now that btrfs is in mainline, let's let anaconda play with
it too  ;)   We still need btrfs.ko in the kernel, and btrfs
support in e2fsprogs.  I've got the latter patch submitted
upstream, and it's in rawhide now.

Note that the below is 100% totally untested and may even have
typos, my track record with python is not too good, sorry.   :)
So review would be appreciated....

Updated to address Radek's review, as well as adding the requisite
witty (commandline-space-eating) boot option...

(And stopped trying to use parted fs identification -- katzj)

Signed-off-by: Eric Sandeen <sandeen@redhat.com>