Strip invalid characters from autoamtically made VG/LV names (#483571).
We make names from mount points and host names, both of which can contain
characters that are invalid to LVM.  The dash is the most promiment case here.
Checks get run when the user enters these bad characters, but not when anaconda
made up the names automatically.  So strip dashes out of the generated names
before feeding them to LVM.