Ensure request.drive is always a list (#485622)
The main problem here is that get_partition_by_name() has been
removed from partedUtils.py.  The functionality is now provided
by the getPartitionByPath() method on a parted.Disk object.  The
old function in partedUtils would iterate over the PedDisk
collection and each partition manually, whereas now we have to
call getPartitionByPath() for each parted.Disk we have.

request.drive has been used inconsistently throughout the existing
storage code and only started to show up when I was moving things
from anaconda to pyparted.  What this patch does is ensure that
request.drive on RequestSpec objects is always a list or None.
There are a few other changes in the storage code to handle the
request.drive variable always being a list, but I think it should
be pretty straightforward.