Fix traceback in timezone setting (kickstart interactive text mode) (#481617)
This is less intrusive (regarding existing kickstart configurations) fix of
bug 404321 - traceback in interactive text mode happening when timezone set in
kickstart is not valid (i.e. not present in zonetab module).  Unlike previous
approach, it doesn't ask user to select valid timezone (breaking unattednded ks
install), but fails to set systemtime for install (if respective zoneinfo file is
not present in stage 2) or configure timezone (if respective zoneinfo file is not
present in tzdata package) only with messages in log files (as it used to be).

Next step is to make valid timezones available for setting of install
environment system time (i.e. present in stage 2).