Include valid timezones in stage 2 (#481617).
Up to now, some valid timezones (zonetab is the reference module)
were missing in stage2, which could cause wrong settig of system time
of install environment in case of hwclock being in localtime.

They are added in the patch while some unused stuff is removed:

1) TIMEZONE variable for specifiyng zoneinfo files from tzdata
to be included in stage2 is removed. Instead of keeping it up-to-date
with zonetab module which would be difficult to do automatically,
whole /usr/share/zoneinfo tree is included which increases size
of stage 2 by about 410 KB.

2) /usr/lib/timezones.gz file is no more generated. The file
contained list of zoneinfo files present in stage2 and had been used
to get list of zones for ui which was replaced with zonetab module.
Hope it is not used anywhere outside anaconda.