Add method pruneActions to remove redundant actions from the queue.
At some point this can probably be triggered from registerAction,
but for now this is simplest. The basic problem is that, when
registering an action, you have to leave the device's previous
action intact so the new action can be cancelled in the event of
a failure like partition allocation. This leads to the possibility
that lots of dialog clicking will lead to a very long sequence of
actions that can/should be distilled to a minimal sequence.

There could be problems lurking here with regard to device
identity/equality.