Fix anaconda udev rules to not require pre-existing device nodes
The way the anaconda udev rules are currently written, they require
the device nodes to already be present when they run. This is not always
the case (as udev creates the nodes after running the rules). This patch
fixes this, which fixes the recognition of lvm VG's and LV's on iscsi disks.