Create partitions with exactly the geometry we calculate.
If we use the device constraint as before, libparted will
align the start and end to a cylinder boundary which, on
my disks, means a granularity of 7MB. Not good enough.

Plus, for mdraid we need >= 1% variation in size between
members of an array. We can get this with cylinder-aligned
partitions, but we'll have to do the alignment ourselves
and continue to use exact constraints.