Speed up partitioning screen redraws by trimming workload where possible.
Previously we would call doPartitioning every time we refreshed
the main partitioning screen. This makes little sense if the
user just finished modifying a vg or md array or only assigned
a mountpoint to an existing device.

There is still more that we could do, like only calling
doPartitioning if one of the generated action involves
partition allocation.