Fix clamping of VG size in case when it is not sufficient for lvm metadata (#489549)
This fixes patch 6c6a206bd8dfc6910fc4061274bd4f054400c838 which
is essentialy kickstart part of a backport from rhel 5.3.
The original patch didn't take in account that PE size is
in KB, not MB. Also 1 PE per PV should be enough for lvm metadata.

I don't port gui part from 5.3 because the bug is for ks and
we are so close to release.