Don't create a PartitionDevice for devices that do not exist (#489122).
This is a strange situation to be in, but there appears to be a corner case
where parted cannot read the disk label despite there being a partition table
on the disk.  anaconda will prompt to create a new disk label.  This means
there's no partitions on the disk, but the list of partitions to probe still
has the old partition list.  So, we should just ignore devices that no longer
exist.