Save a copy of the device stack so we can destroy the format. (#489975)
Since everything is a reference, we are subject to any changes that
may be made to the device/format stack occupied by "our" format from
the time the action is created to the time it is executed. If devices
lower in the stack have their formats scheduled for destruction, we
will be unable to setup the device on which "our" format exists
(because the ActionDestroyFormat constructor unsets the device's
format attribute so the tree can reflect the future state of things).
So, we save a deep copy and are no longer subject to changes
initiated from outside our sphere of interest.