Add mediaPresent and eject to the OpticalDevice class.
These no longer belong in isys.py as putting them in the classes is much
more correct.  I also put a mediaPresent method on Device in general.
This will come in handy when we start dealing with USB CF readers and
similar devices that can be present without having media.