Ignore disk devices with missing media (#488800).
There's a class of device that shows up as a disk drive, but is removable
and doesn't necessarily have media present all the time.  This class
contains things like USB CF readers, those 5-in-1 media readers on the
fronts of some computers, and so forth.

To fix this, I've added a mediaPresent method to DiskDevice that checks if
self.partedDevice was created or not.  We then need to check this throughout
DiskDevice methods - basically, do nothing if we couldn't probe the device
to begin with, but still recognize its existence.

In order to keep the rest of the code clean, we need to weed out all devices
without media present in Storage.disks.