Fix lots of buggy behavior in the partition dialog.
Re-editing new partitions was generating a new request instead of
modifying the original request.

We were also trying to pass in a format to the ActionDestroyFormat
constructor, which is wrong -- they all take a device argument.

Lastly, the conditional for showing the "new" filesystem options
was wrong, which was causing both it and the "preexist" fs options
to get added to the dialog on new request re-edits.