Do not create a LUKSDevice if we do not have a way to map the device.
If the user elects not to provide a passphrase or key, do not create
a LUKSDevice. We will just keep the backing device with its luks
format. This saves us the trouble of checking whether we have the
ability to set up each LUKSDevice since we only have LUKSDevice
instances for those we can set up.