Create a Makefile target to generate updates.img automatically.
Based on what's changed in the repo since the last release tag.
Copy everything necessary for the updates.img in './updates-img'
and pack it up.  Got tired of having my own local scripts to do
this.

LIMITATIONS:  Will not copy in dependent libraries or figure out
if a new _isys.so needs to be built.  Pass KEEP=y if you want the
updates staging directory kept so you can dump other things in to
it.  Run 'make updates KEEP=y' over and over to supplement your
updates image.