Don't use disk.maximizePartition anymore.
We are doing a fine job of maximizing things on our own, and
maximizePartitions is actually shrinking our partitions in an
effort to make sure they are cylinder-aligned. Bottom line,
it's not doing anything useful. If we decide to start making
all partitions cylinder-aligned, we will want to do it in
another way besides calling maximizePartitions, and presumably
our current accuracy will be preserved. If not, we can always
bring it back.