Stop overriding __init__ in DMRaidPartitionDevice
We were overriding __init__ in DMRaidPartitionDevice only to pass
everything through to our supers __init__. Lately we've been doing a bad
job at this as we forget to pass through grow and maxSize causing troubles.

Since all we want to do is pass everything through just not overriding __init__
to begin with is the by far the best fix.