Do not write LV uuid to grub.conf, but the filesystem uuid
When re-using an existing LV for /, and thus one which has uuid set in
its representing Device, we would write the LV uuid to grub.conf as
root= parameter, resulting in a non booting system.

Also we no longer keep labels anywhere, so don't look for a label.

Note that we are only hitting this with pre-existing LV's because we
are not setting / updating the LVDevice's uuid when creating a new one
(something which we ought to fix).