Attempt disk commits 5 times before raising an exception.
Hitting this periodically on my test systems.  The disk may or may
not be capable of the ped_disk_commit_to_dev() call yet.  The disk
is busy.  I don't want to use a Timer thread since we're using
PyGTK and we'll get in to all sorts of problems if threads are
brought it.  The poor man's try loop has been working for me today.