Do not use _rnetdev as fstab option for network based /
In RHEL-5 we use _rnetdev for network based /, and various pieces
of the RHEL-5 initscripts key of on this to do special things. So I thought
it would be good to do the same in Fedora. Testing has revealed that
Fedora's mount command cannot handle this options on things go boom,
so use plain _netdev for / just like we do with other network based fs, and
as we've been doing in previous Fedora releases.