Be a little bit smarter about allocating space to grow parts. (#491761)
parted.disk.Disk.maximizePartition leads to libparted applying
a cylinder-alignment constraint, which we would frikking ask for
if we wanted it. So, for now, we can use this to identify chunks
of free space that will initially be allocated to requests whose
max size will prevent them from making full use of their share
of the space.

Cases where this comes into play are ones where an initially
larger request has a relatively low max size and an initially
smaller request has no max size at all. Since we base the
allocation of free space for growing on the requests' initial
size, we could previously end up with very large chunks of
unused free space.