Add an early kickstart processing pass.
This gets rid of the old several early passes for VNC and rescue, instead
replacing them with a single one that also looks for ignoredisks.  We need
to look for ignoredisks early on so that we have that list ready before
enabling storage.  Otherwise, ignoredisks won't really do what it's supposed
to.  The problem here is what to do about if the ignoredisks directive is
located in an include generated by a %pre script.