Correct bounds checking problems in 'Shrink current system'
The 'Shrink current system' dialog had some bounds checking problems.
Users could enter a size that was too small or too large.

Renamed the function and related controls to use 'shrink' instead of
'resize' in the name to better reflect what the code does.

Use the format's size to set the bounds rather than the partition's.
Since we are shrinking here, we only want to let users select a size
smaller than the current format size.

A size smaller than the minimum will automatically be set to the correct
minimum.  Likewise, a size larger than the maximum will automatically be
set to the correct maximum.