Do not use getPartitionByPath() in allocatePartitions()
parted always renumbers partitions when a partition gets removed, so
if we have p5 and p6, and then p5 gets removed then p6 becomes p5, if
we then check for p6 using getPartitionByPath() it will show up as
non existing even though it does exist, but it has been renamed to p5.

This patch fixes this by simply setting the PartitionDevice's pedPartition
to None when it was removed from its disk, as the PartitionDevice is of no
use to use after that anyways (and we already check that we do not try
to remove None from the disk).