Don't exit the installer from filesystem errors.
We shouldn't be mounting any filesystems in DeviceTree.populate. The
only way we should be able to get an FSError is doing the minSize
calculation, which should not raise exceptions in the first place. If
we do have a way to get filesystem errors, let's find out what they are
and fix them instead of masking them.