Preserve symlinks and only collect deps on ELF executables.
The mount and umount commands will have a bunch of symlinks, so if the
command we are installing is a symlink, preserve it.

Only use -s on install(1) if the command we are installing is an ELF
executable.  Since there's a check for that, use it to fire off
get_dso_deps() for only ELF executables.