Pre-existing partitions names may change (#494833)
When there are multiple (pre-existing) logical partitions and a logical
partition other then the last partition gets deleted, the number
of the partitions after it changes. Causing getDeviceByName, with the
name based on the new partition number (gotten through parted) to fail.

Specifically this causes a "'NoneType' object has no attribute 'path'"
exception in populate() in partition_gui.py . This patch fixes this by
updating all partitions names in doPartitioning.

Note that for this patch to help the following pyparted bug must also
be fixed: https://fedorahosted.org/pyparted/ticket/17