Fix display of format type for devices.
When re-editing a new device, the fstype combo was not getting set to
the current format type because of confusion caused by the use of
format name in some places and format type in others.

The only thing we use the name attribute for is displaying in the
fstype combo. For everything else we use the type attribute.