growPartitions: Change op_func (back to) add when an iteration succeeds
Currently when an iteration fails we will switch from growing to
shrinking our req_size, but we neglect to switch back to growing when we
succeed. So if for example orig_sectors + 1/2 max_grow would fail, we
would try orig_sectors + 1/4 max_grow and even if that succeeds we would
then try orig_sectors + 1/8 max_grow (and keep that) instead of trying
orig_sectors + 3/8 max_grow.