Determine minimum size for filesystems once, from constructor.
This way we won't have to setup complex devices every time the
maxSize attribute is accessed. The device's minimum size should
never change without the user doing something that triggers a
storage reset/rescan, so there is no reason to expect that caching
the minimum size will lead to new failures. We know that the device
is already setup when the constructor is called, so it is a perfect
time to obtain the minimum size and be done with it.