Don't traceback on invalid filesystem detection (#495156)
We may think a partition has a valid filesystem when it really does not.
This is easily reproduced with the following steps:

1) Create a partition on some device.
2) Create an NTFS filesystem on that partition.
3) Delete the partition from the device.
4) Divide the device in to two equal sized partitions, but do not
   create a filesystem on either.

Boot the installer and notice that the first partition is assumed to be
NTFS, but it's unmountable.  In getExistingSize(), we get an FSError
traceback when trying to mount the filesyste.

The problem is that we haven't zeroed out the new partition or created a
new filesystem on top of it.  The old filesystem data is still on the
disk and is misleading udev and friends.

The solution in this patch is to catch the FSError from
getExistingSize().  If the mount failed, assume the partition is empty
and continue.