use partition req_base_size instead of size in partitionCompare()
Currently we are using part.size in partitionCompare(), this will
return the grown size of a partition if it is already grown at this
point, potentially changing the order in which we allocate partitions
while we are growing. By using req_base_size we will always allocate
the partitions in the same order while growing multiple partitions.