Bring luks_add_key and luks_remove_key back into devicelibs.crypto.
In order to include these in pycryptsetup we had to create yet
another duplicate copy of execWithRedirect. This patch brings them
back into anaconda so we can easily modify them if needed until
the day when pycryptsetup has this functionality. It also allows us
to redirect the output to someplace useful for debugging.