Use size instead of currentSize when comparing lv sizes
parted and lvm have a different idea of the exact size of an LV, so
when using currentSize to see if we need to resize, we will always
create a ResizeAction, which will later on fail as "lvm lvresize"
returns an error status if the requested resize is a nop.

So use size instead, this is safe as we never change the size of
existing lv's.

This fixes a traceback with a lvmresize exception when trying to use
an existing LV during install without changing its size.