Clean up argument list after changing from rhpl to iutil for execWithRedirect
iutil uses subprocess.Popen and passes in the argv as a list, which causes
Popen to use the first entry of the list as the executable, and the rest of
the entires as arguments to that executable.  This is different from how
rhpl did things in which the first argument of the arglist had to be the
executable itself.

Also iutil expects argv to be a real list not a tuple.