Fix handling of bind mounts. (#496406)
When we're parsing /etc/fstab the directories that serve as the devices
in bind mounts are not going to be present, so we will have a hell of a
time figuring out what devices contain them. However, in
FSSet.mountFilesystems we should have those dirs set up already, making
it possible to sort out.