Don't remove an inconsistent lvm partition from the devicetree (#496638)
When a PV partition is part of an inconsistent VG, and the user chooses
to ignore it we currently remove it from the devicetree, but since
partition_gui.py populate() uses parted on the disk to find out about
the partitions, it will still see it and then backtrace when it cannot
find it in the devicetree.

This patch fixes this by instead making the partition immutable. Note
that I've put the reason for it being immutable in the new immutableDevices
array, so that if we have similar cases in the future we can use
immutableDevices for that too. We might even want to move some of our
existing cases there.