Properly restore SIGCHLD if X startup fails
Before starting X we set up a SIGCHLD handler to raise OSError if
Xorg dies. But if that happens, we'll skip over the code that
restores the old SIGCHLD handler, and thus the next subprocess we
run will cause anaconda to catch OSError and die.

This patch moves the signal restore code into a finally: block.