Fix traceback in second editing of existing raid dev (#497234).
This only fixes the traceback, but not the deeper cause which
is that we are not editing original format request, but our last
request (edit), so e.g. in 2nd edit "Original filesystem type"
label and "Migrate" option are wrong (reflecting our last
request, not the original format). I added some more conditions
so that we do not display the info and do not offer migration
in such cases. Note that the checks also affect non-raid
existing partition editing in the same sense. The conditions
should be kept when the behavior is fixed.