Use storage objects throughout the partition editing UI code (#491806, #496002).
This fixes the problem where highlighting the partition in the slice graph
doesn't allow you to click "edit", and where selecting from the tree view
doesn't highlight the partition in the graph.  The underlying problem was
that our list stores storage objects, but we were passing in a pyparted
object to do the comparison with.