Add a Mac OS boot line to yaboot.conf (#499964).
The problem is that a Mac filesystem is HFS+, but is not marked bootable.
The Apple Bootstrap partition is marked bootable.  Therefore we take the
stupid test and say if there's a bootable Apple Bootstrap partition on a
disk and there's an HFS partition, mark that as the Mac OS partition for
the bootloader config.  This is about as stupid as what we're doing for
DOS filesystems now and is easily screwed up, but it's better than
nothing.