Consider active-idle state of md device as accepatable status of device (#497407)
This should fix traceback met when we were updating sysfs path attribute of md
device in active-idle state (empty string was returned as the device was not in
state clean or active). I think it is a state very rarely met in our tests.
According to md.txt, it is "like active, but no writes have been seen for a
while".  I prepared updates.img for testing, but we can't reproduce the bug.