Do not allow users to migrate ext4 to ext4.
Noticed that ext4 filesystems were presenting users with a 'migrate to'
combo box with 'ext4' as the only entry.  Tracked down the problem to
the migratable property override in Ext3FS.  Changed it around to use
helper functions with property() and the UI is working more sanely now.
Users cannot select ext4 as the migration target for ext4.