Remove LVM metadata when doing clearpart (#462615)
LVM metadata are removed 1) when labeling a disk (covers clearpart --initlabel
cases) or 2) when deleting a partition upon a delete request if the request was
created in clearpart.  Also we don't create dependent meta delete requests in
this case as they would conflict with metadata removing in partition delete.
If the delete request comes from UI, LVM metadata are removed as before by meta
delete (VG) requests created in UI in previous steps.