Fix detection of cciss disks (#499408)
Since cciss device names have a subdirectory (i.e. cciss/c0d0p1) they are
represented in sysfs with a ! so that it doesn't incorrectly have an extra
subdir in the sysfs namespace.  When anaconda gets the device name from sysfs
it was trying to lookup the name with a ! but anaconda stores them with the /
in the name.