We are not guaranteed to have a partedDisk in the udev code (#501556, #501428).
This goes back to having drives with no media present, like USB media readers.
At this point in the code, we can't yet use mediaPresent because we don't have
all the device objects set up.  So, we just have to rely on partedDisk being
None.  Besides, checking for None is always a good idea anyway.