Make the save-to-bugzilla dupe detection smarter.
This fixes two problems with how we determine tracebacks are the same based
on filename comparisons:

(1) Comparisons between python files on i386 and x86-64 tracebacks fail because
/usr/lib and /usr/lib64 are different, even though the files themselves are the
same.

(2) Comparisons between python files on the running system and in /tmp/updates
fail obviously.  This is a little less concerning to fail on, but since the
lines of code are likely to change in an updates.img, duplicates should still
work properly.