Use gettext.ldngettext when necessary (#467603)
The i18n people have suggested using ngettext when we need to have
singular and plural forms of strings, where the count will vary as to
what we are reporting to the user.

I've made the changes they have suggested.  I created a new lambda
function called P_() to use for the plural cases.  P_() takes in three
parameters:

    1) The singular form of the string.
    2) The plural form of the string.
    3) A count.

Here's an example:

    ....some loop runs doing stuff
    bytesWritten = 47
    msg = P_("Wrote %d byte.",
             "Wrote %d bytes.",
             bytesWritten) % (bytesWritten,)
    print msg

The % substitution is correct at the end because P_() returns a single
string, so we only need the format string to account for that.

Some strings have been changed slightly to make it easier for
translations to other languages, particularly when choosing plural forms.