Test NM_CONTROLLED setting correctly in network.py (#502466)
The only value we care about for NM_CONTROLLED is 'no', so change test
to work that way.  If performing a livecd install, make sure to carry
over any NM_CONTROLLED settings in the ifcfg-* files on the live CD.
Users with custom live images may have that set, so we should preserve
it.