Restructure liveinst/ for the new build system.
Added liveinst/Makefile.am and restructured the subdirectory to work a
little more easily with autoconf and automake.  The liveinst.console
file has been moved to console.apps/liveinst.  The liveinst.pam file has
been moved to pam.d/liveinst.  Additional Makefile.am files are present
in those subdirectories.

The restructure of liveinst/ is to account for the fact that most of the
files we have to install are named 'liveinst'.