Convert po/ subdirectory to GNU gettext template system.
Maintaining po/ now means maintaining po/LINGUAS, po/POTFILES.in, and
po/Makevars.  The other files are automatically added by gettextize when
we update to a newer minimum version of gettext.

The LINGUAS file lists all of the languages we support.  The POTFILES.in
file lists all of the source files that have strings for translation.
The Makevars file sets some control variables and options to xgettext.