Rework tracking of devices containing installation media. (#497087)
In storage.Storage we keep the device specifications as provided to us.
In the devicetree, during population, we use udev to resolve the user-
provided specs to device names. As we create StorageDevice instances for
the devices, we set a new attribute ("protected") as appropriate. Once
the DeviceTree is populated, we use the devices' protected attribute to
determine whether or not they contain installation media. This way we
can track protected devices even when their names change.