no "rhgb quiet" on s390 to enable visible boot progress and system automation (#509881)
Anaconda appends "rhgb quiet" to the bootloader's kernel arguments.
On s390x, there is not graphical console. A console of type line mode is most
common. Even more so, it's not uncommon to do system automation based on the
guest OS's console output. Therefore, allow full output of kernel boot messages on the console.