Indicate LV status according to lv_attr active bit (#491754)
This should fix up a majority, if not all, of the 'cannot commit to disk
after 5 attempts' errors.  I was hitting this today.  The cause I found
to be active logical volumes still around when the storage code wanted
to commit changes to the disk.  When a logical volume is active,
libparted is getting EBUSY when it tries to commit changes to the disk
and tell the kernel to reread the partition table.  If the LV is active,
we need to deactivate it and the volume group before we start committing
changes to disk.

Adding a status() property to LVMLogicalVolumeDevice that looks at the
lv_attr field for an 'a' fixes the problem for me on F-11.  All of the
code to down LVM devices is there, we just weren't checking the LV
status correctly.