70-anaconda.rules: always import blkid output
We are currently not calling blkid when a block device is not an
mdraid set or a device mapper device, resulting in us not identifying
mdraid members as such. This probably worked when still using vol_id
because some other rules file was doing an import vol_id before our rules.