Reverse: "Support for MD containers"
Reverse most of commit 9e4d2e76713c9c71dcdaf22767211c13be9b3668, because it
currently is not working and MD containers can for 99.9 % be treated as
regular mdarrays. The next patch from this set instead makes the necessary
changes to use the regular mdarray handling code, reducing the amount
of code.