Call mdadm -I with --no-degraded for all disks but the last
We do not want to bring arrays up (and esp not use, so not scan them) in
degraded mode, so add --no-degraded to the mdadm invocation.

Note we do not add --no-degraded to the mdadm invocation for the last
disk of the set, because doing so assembles the array in an inactive
(and thus unusable) state.