70-anaconda.rules: pass --ignorelockingfailure to lvm invocation
During testing I noticed that anaconda did not recognize pre-existing
logical volumes and thus that I could not upgrade / rescue an existing
default partitioning install.

I tracked this down to the last lvm call in 70-anaconda.rules failing
(udevadm test is your friend). Passing --ignorelockingfailure to lvm
fixes this. For good measure this patch does this for all invocations in
70-anaconda.rules