Adapt exception.py to using python-meh.
This involves getting rid of most of the real code out of this file, adding a
minimal subclass of the ExceptionHandler class to do the copying and pdb setup
anaconda needs, and calling the install method from within anaconda.  We also
need to no longer call handleException from elsewhere in anaconda.  That method
no longer exists.