Add scripts/makeupdates to generate updates.img files.
Move the contents of the 'make updates' target to scripts/makeupdates.
Make the recompilation of the _isys module be conditional (if you want
to do that, either run 'make updates' as you have been or pass -c to
scripts/makeupdates).  Running makeupdates without -c will generate an
updates.img with just the changed Python files.