Edit LVM LV when user has a LV selected.
* iw/lvm_dialog_gui.py (run, convertToActions): New function.  Need to
have a separate function so we can run it from the editLVMLogicalVolume
function in iw/partition_gui.py.  We call new function from run.
* iw/partition_gui.py (editCB): Call the new editLVMLogicalVolume
function.
(editLVMLogicalVolume):  New function.  Serves the same purpose as
editLVMVolumeGroup but for Logical Volumes.
(deleteCB): All actions are created when we delete a selected LV.  We do
need to make sure that we delete the deleted LV from the VG that