Work around problems with live installs and dpi other than 96 (#506512)
On the live install, we don't force the DPI to 96.  Some of our text
takes a bit of space, especially the autopart screen.  So if we're on a live
install (eg, without mini-wm running), let's just always get rid of the
little header image.  It doesn't really buy us much and then we
get more real estate to work with.

Longer term, we probably need to move away from the fixed window
size stuff, even though it'll be painful