Do not offer going back when ugrade root for ks upgrade is not found (#499321)
Going back results in traceback which seems not right. Bug 499321 (preupgrade
backtrace) has lots of tracebacks with various causes of upgrade root not found
- I distincted two groups.:  1) dmraid not detected in f10 (=> installation on
2 disks) while detected in f11 (=> root not found). 2) root considered as
non-upgradable (/etc/redhat-release changed by user). Such a broad bug is
rather difficult to handle. OTOH we wouldn't get the tracebacks so easy if we
just ended with error message (as with the patch).