Include ipcalc command in all initrd.img files, not just s390 (#516084)
If a user runs dhclient by hand from the installer media (say, if they
are using the installer media to just bring up a shell with a Linux
environment), they will get the following error:

    /sbin/dhclient-script: line 384: ipcalc: command not found

We have been including ipcalc in the s390 initrd.img, because
linuxrc.s390 requires it.  Move the inclusion in upd-instroot to all
architectures instead.