Do not go interactive if timezone in ks is not valid (#473647)
Handle invalid timezones in ks the same way as in RHEL 5.4:
in case that the timezone is invalid, the system time of install environment
will be set correctly if it is in tzdata package, and even if it is not there
the ks install will proceed.  This will prevent possible breaking of installs
with kickstarts working in older versions of RHEL.

The question may be if we can/want to broaden valid timezones to all those
present in tzdata package officialy (document it). Complete tzdata package
is part of stage2 in Fedora and will be in RHEL 5.4.