Honor network config boot params for CD-booted ks installs (#433214)
Users booting from CD or DVD media, performing a kickstart install, and
passing network configuration boot parameters were hitting an issue
where loader was insisting on trying to get a DHCP lease to fetch the
kickstart file.  Even if you provided 'asknetwork' or 'ip=' settings,
loader was still forcing DHCP at the point where the kickstart file is
downloaded.

This patch makes the 'asknetwork' parameter work if you are booting from
CD or DVD and performing a kickstart install.  By default, we would try
DHCP, but asknetwork forces the configureTCPIP() code path to run.

Likewise, if the user boots from media and passes ip=, netmask=,
gateway=, dns=, and ksdevice=, the network is configured using that
information and the kickstart file is downloaded.