Add support for the reiserfs filesystem (#504401)
As the bug indicates, the kernel module is still available.  I'm not
really interested in supporting installation to reiserfs, but I think we
should be able to recognize the filesystems in the partitioning UI as
well as let people shrink existing reiserfs filesystems to make room for
Fedora on more sane filesystems.

Detection works as is, but users have to pass 'reiserfs' on the command
line to enable installation to reiserfs volumes or to enable resize
support.