Fix focus grabbing on both the password and hostname screens.
Since we have the Next button grab focus immediately before drawing then
screens, we'd been relying on a complicated system of GTK signals to do
the job but that only ever worked part of the time.  Instead, add a focus()
method to InstallWindow that screens can override and will set the focus
to the correct widget.