disklabel.commit(): DeviceError -> DeviceFormatError
While running anaconda with a patched libparted, which will actually return
an error upon commiting to a busy disk I got a traceback as the packages.py
exception handling tried to unpack the disklabel.commit() exception into
a message and a device, as disklabel.commit() was throwing a DeviceError
which normal has a message and a device. Since disklabel is a format now
commit() should throw a DeviceFormatError instead.

(And now to find out why the commit is failing)