Do not try to commit disks changes to the os while partitions are in use
In storage.doIt() we set the bootable flag for the /boot or / (or
special platform boot) partition. And then we call bootdev.format.commit(),
however this will lead to parted trying to tell the kernel to rescan the
partition table which will usually fail, as usually we have some VolGroup
active at this point.

All we really need todo is make sure the bootable flag gets written to the
disk, so this patch adds a commitToDisk() method to formats.disklabel and uses
that.

We were not bitten by this before because uptil now parted was silently
ignoring ebusy errors for us (BAD parted, BAD!).